+++ lib/src/i69n_impl.dart
@@ -1,6 +1,7 @@
 library i69n;
 
 import 'package:yaml/yaml.dart';
+import 'package:dart_style/dart_style.dart';
 
 part 'model.dart';
 
@@ -25,26 +26,20 @@ String generateDartContentFromYaml(ClassMeta meta, String yamlContent) {
   output.writeln("String get _languageCode => '${meta.languageCode}';");
   output.writeln("String get _localeName => '${meta.localeName}';");
   output.writeln("");
-  output.writeln(
-      "String _plural(int count, {String zero, String one, String two, String few, String many, String other}) =>");
-  output.writeln(
-      "\ti69n.plural(count, _languageCode, zero:zero, one:one, two:two, few:few, many:many, other:other);");
-  output.writeln(
-      "String _ordinal(int count, {String zero, String one, String two, String few, String many, String other}) =>");
-  output.writeln(
-      "\ti69n.ordinal(count, _languageCode, zero:zero, one:one, two:two, few:few, many:many, other:other);");
-  output.writeln(
-      "String _cardinal(int count, {String zero, String one, String two, String few, String many, String other}) =>");
-  output.writeln(
-      "\ti69n.cardinal(count, _languageCode, zero:zero, one:one, two:two, few:few, many:many, other:other);");
+  output.writeln("String _plural(int count, {String zero, String one, String two, String few, String many, String other}) =>");
+  output.writeln("\ti69n.plural(count, _languageCode, zero:zero, one:one, two:two, few:few, many:many, other:other);");
+  output.writeln("String _ordinal(int count, {String zero, String one, String two, String few, String many, String other}) =>");
+  output.writeln("\ti69n.ordinal(count, _languageCode, zero:zero, one:one, two:two, few:few, many:many, other:other);");
+  output.writeln("String _cardinal(int count, {String zero, String one, String two, String few, String many, String other}) =>");
+  output.writeln("\ti69n.cardinal(count, _languageCode, zero:zero, one:one, two:two, few:few, many:many, other:other);");
   output.writeln("");
 
   for (var todo in todoList) {
     renderTodoItem(todo, output);
     output.writeln("");
   }
-
-  return output.toString();
+  var formatter = DartFormatter();
+  return formatter.format(output.toString());
 }
 
 ClassMeta generateMessageObjectName(String fileName) {
@@ -75,8 +70,7 @@ ClassMeta generateMessageObjectName(String fileName) {
     if (nameParts.length >= 2) {
       String languageCode = nameParts[1];
       if (twoCharsLower.allMatches(languageCode).length != 1) {
-        throw new Exception(
-            "Wrong language code '$languageCode' in file name '$fileName'. Language code must match $twoCharsLower");
+        throw new Exception("Wrong language code '$languageCode' in file name '$fileName'. Language code must match $twoCharsLower");
       }
       result.languageCode = languageCode;
       result.localeName = languageCode;
@@ -84,8 +78,7 @@ ClassMeta generateMessageObjectName(String fileName) {
     if (nameParts.length == 3) {
       String countryCode = nameParts[2];
       if (twoCharsUpper.allMatches(countryCode).length != 1) {
-        throw new Exception(
-            "Wrong country code '$countryCode' in file name '$fileName'. Country code must match $twoCharsUpper");
+        throw new Exception("Wrong country code '$countryCode' in file name '$fileName'. Country code must match $twoCharsUpper");
       }
       result.localeName = "${result.languageCode}_${countryCode}";
     }
@@ -98,10 +91,9 @@ void renderTodoItem(TodoItem todo, StringBuffer output) {
   var meta = todo.meta;
   YamlMap content = todo.content;
   if (meta.isDefault) {
-    output.writeln("class ${meta.objectName} {");
+    output.writeln("class ${meta.objectName} implements i69n.I69nMessageBundle {");
   } else {
-    output.writeln(
-        "class ${meta.objectName} extends ${meta.defaultObjectName} {");
+    output.writeln("class ${meta.objectName} extends ${meta.defaultObjectName} {");
   }
 
   if (meta.parent == null) {
@@ -111,30 +103,51 @@ void renderTodoItem(TodoItem todo, StringBuffer output) {
     if (meta.isDefault) {
       output.writeln("\tconst ${meta.objectName}(this._parent);");
     } else {
-      output
-          .writeln("\tconst ${meta.objectName}(this._parent):super(_parent);");
+      output.writeln("\tconst ${meta.objectName}(this._parent):super(_parent);");
     }
   }
   content.forEach((k, v) {
     if (v is YamlMap) {
       String prefix = _firstCharUpper(k);
       ClassMeta child = meta.nest(prefix);
-      output.writeln(
-          "\t${child.objectName} get ${k} => ${child.objectName}(this);");
+      output.writeln("\t${child.objectName} get ${k} => ${child.objectName}(this);");
     } else {
       if (k.contains("(")) {
         // function
         output.writeln('\tString ${k} => "${v}";');
       } else {
+        if (k.contains(".")) {
+          throw Exception("Your message key cannot contain a dot, see '$k'");
+        }
         output.writeln('\tString get ${k} => "${v}";');
       }
     }
   });
+  output.writeln("\tObject operator[](String key) {");
+  output.writeln("\t\tint index = key.indexOf('.');");
+  output.writeln("\t\tif (index > 0) {");
+  output.writeln("\t\t\treturn (this[key.substring(0,index)] as i69n.I69nMessageBundle)[key.substring(index+1)];");
+  output.writeln("\t\t}");
+  output.writeln("\t\tswitch(key) {");
+  content.forEach((k, v) {
+    String key = k;
+    if (key.contains('(')) {
+      key = key.substring(0, key.indexOf('('));
+    }
+    output.writeln("\t\t\tcase '$key': return $key;");
+  });
+  if (meta.isDefault) {
+    output.writeln("\t\t\tdefault: throw new Exception(\"Message \'\$key\' doesn\'t exist in \$this\");");
+  } else {
+    output.writeln("\t\t\tdefault: return super[key];");
+  }
+  output.writeln("\t\t}");
+  output.writeln("\t}");
+
   output.writeln("}");
 }
 
-void prepareTodoList(
-    List<TodoItem> todoList, YamlMap messages, ClassMeta name) {
+void prepareTodoList(List<TodoItem> todoList, YamlMap messages, ClassMeta name) {
   TodoItem todo = new TodoItem(name, messages);
   todoList.add(todo);